requirements.txt
black==21.9b0
coreapi==2.3.3
coreschema==0.0.4
Django==3.2.7
django-common-helpers==0.9.2
django-cors-headers==3.9.0
django-cron==0.5.1
django-crontab==0.7.1
djangorestframework==3.12.4
djangorestframework-simplejwt==4.8.0
djoser==2.1.0
drf-yasg==1.20.0
flake8==3.9.2
oauthlib==3.1.1
openapi-codec==1.3.2
psycopg2-binary==2.9.1
pytz==2021.3